app/requirements.txt
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
aiohttp==3.13.2